requirements.txt
numpy
gymnasium
gym-retro
opencv-python
scikit-image
torch>=2.0.0
torchvision
torchaudio